fix: wasm events empty on logs
