+++ relayer/chains/wasm/types/types.go
@@ -54,13 +54,28 @@ func (param *TxSearchParam) BuildQuery() string {
 	return finalQuery.GetQuery()
 }
 
-type TxResultWaitResponse struct {
+type Event struct {
+	Type       string      `json:"type"`
+	Attributes []Attribute `json:"attributes"`
+}
+
+type Attribute struct {
+	Key   string `json:"key"`
+	Value string `json:"value"`
+}
+
+type EventsList struct {
+	Events []Event `json:"events"`
+}
+
+type TxResultResponse struct {
 	Height int64 `json:"height"`
 	Result struct {
-		Code      int    `json:"code"`
-		Codespace string `json:"codespace"`
-		Data      []byte `json:"data"`
-		Log       string `json:"log"`
+		Code      int     `json:"code"`
+		Codespace string  `json:"codespace"`
+		Data      []byte  `json:"data"`
+		Log       string  `json:"log"`
+		Events    []Event `json:"events"`
 	} `json:"result"`
 }
